requirements.txt
attrs==23.2.0
beautifulsoup4==4.11.1
blinker==1.7.0
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.1.1
cryptography==42.0.8
exceptiongroup==1.2.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
kaitaistruct==0.10
outcome==1.3.0.post0
pyasn1==0.6.0
pycparser==2.22
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
PyYAML==6.0
requests==2.32.3
requests-toolbelt==0.10.1
selenium==4.21.0
selenium-stealth==1.0.6
selenium-wire==5.1.0
slack-sdk==3.19.5
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
wsproto==1.2.0
zstandard==0.22.0
